+++ src/assets/icons/navigationIcon.js
