+++ src/components/modal/Modal.jsx
@@ -1,5 +1,5 @@
 import Picker from "emoji-picker-react";
-import { Button, PostImageButton } from "pages/feeds/feedsComponent";
+import { Button, PostImageButton } from "features/home/pages/feedsComponent";
 import {
   ModalWrapper,
   CloseButton,
@@ -9,11 +9,11 @@ import {
   PickerWrapper,
   InputImage
 } from "./modalComponent";
-import addPostImg from "Assets/images/addpostImg.webp";
+import addPostImg from "assets/images/addpostImg.webp";
 import { useDispatch } from "react-redux";
 import { useState, useEffect } from "react";
 import { AlertToast } from "../toasts";
-import { addNewPost, editPost,setEditEmpty,setModalOpen } from "Redux/Reducers/postsSlice";
+import { addNewPost, editPost,setEditEmpty,setModalOpen } from "features/home/postsSlice";
 import { getBgColor, getTextColor } from "utils/Functions/getColor";
 import { useTheme } from "context/theme-context";
 import { useSelector } from "react-redux";
