+++ src/features/authentication/pages/SignUp.jsx
@@ -9,7 +9,7 @@ import {
 } from "./AuthFormComponent";
 import { useState } from "react";
 import {useDispatch} from "react-redux"
-import { signup } from "Redux/Reducers/authSlice";
+import { signup } from "features/authentication/authSlice";
 import {AlertToast} from "components/toasts"
 import {useDocumentTitle} from "utils/hooks/useDocumentTitle";
