+++ src/features/index.js
@@ -1,0 +1,9 @@
+export {Login} from "./authentication/pages/Login"
+export {SignUp} from "./authentication/pages/SignUp"
+export {Explore} from "./explore/Explore"
+export {Profile} from "./user/pages/profile/Profile"
+export {Setting} from "./user/pages/setting/Setting"
+export {Landing} from "./landing/pages/Landing"
+export {Feeds} from "./home/pages/Feeds"
+export {SinglePost} from "./singlepost/pages/SinglePost"
+
