+++ src/features/landing/pages/Landing.jsx
@@ -1,6 +1,6 @@
 import {Wrapper,Section,HeroImg,Header,Para,Button} from "./landingComponents"
-import landinheroImg from "Assets/images/landingheroImg.png"
-import {Logo} from "Assets/icons/Logo"
+import landinheroImg from "assets/images/landingheroImg.png"
+import {Logo} from "assets/icons/Logo"
 import {useDocumentTitle} from "utils/hooks/useDocumentTitle";
 export const Landing = () =>{
     useDocumentTitle("Landing")
