+++ src/features/user/pages/profile/Profile.jsx
@@ -12,7 +12,7 @@ import {
 } from "./profileComponents";
 import { getBorder } from "utils/Functions/getColor";
 import { useTheme } from "context/theme-context";
-import { Tabs } from "./Tab/Tabs";
+import { Tabs } from "./tab/Tabs";
 import { useNavigate, useParams } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { useEffect, useState } from "react";
