+++ src/index.js
@@ -1,11 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "App";
+import App from "app/App";
 import { makeServer } from "server";
 import { ThemeProvider } from "context/theme-context";
 import { BrowserRouter } from "react-router-dom";
 import {Provider} from "react-redux"
-import store  from "Redux/store"
+import store  from "app/store"
 // Call make Server
 makeServer();
