--- src/pages/index.js
@@ -1,8 +1,0 @@
-export {Landing} from "./Landing/Landing"
-export {Feeds} from "./feeds/Feeds"
-export {Login} from "./Authpages/Login"
-export {SignUp} from "./Authpages/SignUp"
-export {Profile} from "./Profile/Profile"
-export {Setting} from "./Setting/Setting"
-export {Explore} from "./Explore/Explore"
-export {SinglePost} from "./singlePost/SinglePost"
